GeoCode response class

+++ src/main/java/hello/ShopFindController.java
@@ -4,8 +4,6 @@ import hello.model.Greeting;
 import hello.model.ShopDetailsRequest;
 import hello.model.ShopDetailsResponse;
 
-import java.math.BigInteger;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -23,8 +21,8 @@ public interface ShopFindController {
     public ResponseEntity<ShopDetailsResponse> addShop(ShopDetailsRequest request);
 
     @RequestMapping(method = RequestMethod.GET, value = "/getShop")
-    public ResponseEntity<ShopDetailsResponse> getShop(@RequestParam(value = "latitude") BigInteger latitude,
-            @RequestParam(value = "longitude") BigInteger longitude);
+    public ResponseEntity<ShopDetailsResponse> getShop(@RequestParam(value = "latitude") Double latitude,
+            @RequestParam(value = "longitude") Double longitude);
 
     @RequestMapping("/greeting")
     public Greeting greeting(@RequestParam(value = "name", defaultValue = "World") String name);
